Refactor: move UserService interface & UserServiceImpl to correct packages
